requirements.txt
pandas
python-dotenv
tqdm
langchain-openai
langchain-core
openai